add untracked changes

+++ Infrastructure/Repositories/ExamRepository.cs
@@ -7,6 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using static System.Runtime.InteropServices.JavaScript.JSType;
 
 namespace Infrastructure.Repositories
 {
@@ -14,16 +15,22 @@ namespace Infrastructure.Repositories
     {
         private readonly AppDbContext _db;
         public ExamRepository(AppDbContext db) => _db = db;
-        public async Task<Exam?> GetByIdAsync(int id, CancellationToken ct = default)
-        => await _db.Exams.FirstOrDefaultAsync(x => x.ID == id, ct);
+        public async Task<Exam?> GetByKeyAsync(int studentId, int subjectId, DateOnly date, CancellationToken ct = default)
+        => await _db.Exams.FirstOrDefaultAsync(x =>
+            x.StudentID == studentId &&
+            x.SubjectID == subjectId &&
+            x.Date == date, ct);
 
-        public Task<Exam?> GetByIdWithDetailsAsync(int id, CancellationToken ct = default)
+        public Task<Exam?> GetByKeyWithDetailsAsync(int studentId, int subjectId, DateOnly date, CancellationToken ct = default)
             => _db.Exams
-                  .Include(x => x.Student)
-                  .Include(x => x.Subject)
-                  .Include(x => x.Examiner)
-                  .Include(x => x.Supervisor)
-                  .FirstOrDefaultAsync(x => x.ID == id, ct);
+            .Include(x => x.Student)
+            .Include(x => x.Subject)
+            .Include(x => x.Examiner)
+            .Include(x => x.Supervisor)
+            .FirstOrDefaultAsync(e =>
+               e.StudentID == studentId &&
+               e.SubjectID == subjectId &&
+               e.Date == date, ct);
 
         public async Task<IReadOnlyList<Exam>> ListWithDetailsAsync(CancellationToken ct = default)
             => await _db.Exams
@@ -35,11 +42,10 @@ namespace Infrastructure.Repositories
                 .OrderByDescending(x => x.Date).ThenBy(x => x.StudentID)
                 .ToListAsync(ct);
 
-        public async Task<int> CreateAsync(Exam exam, CancellationToken ct = default)
+        public async Task CreateAsync(Exam exam, CancellationToken ct = default)
         {
             _db.Exams.Add(exam);
             await _db.SaveChangesAsync(ct);
-            return exam.ID;
         }
 
         public async Task UpdateAsync(Exam exam, CancellationToken ct = default)
@@ -47,24 +53,25 @@ namespace Infrastructure.Repositories
             await _db.SaveChangesAsync(ct);
         }
 
-        public async Task DeleteAsync(Exam exam, CancellationToken ct = default)
+        public async Task DeleteAsync(int studentId, int subjectId, DateOnly date, CancellationToken ct = default)
         {
+            var exam = await GetByKeyAsync(studentId, subjectId, date, ct);
+            if (exam is null) return;
             _db.Exams.Remove(exam);
             await _db.SaveChangesAsync(ct);
         }
-
         public async Task<bool> ExistsOnDateAsync(int studentId, int subjectId, DateOnly date, CancellationToken ct = default)
         {
-           return await _db.Exams.AnyAsync(e => e.StudentID == studentId
-                                  && e.SubjectID == subjectId
-                                  && e.Date == date, ct);
+            return await _db.Exams.AnyAsync(e =>
+             e.StudentID == studentId &&
+             e.SubjectID == subjectId &&
+             e.Date == date, ct);
         }
-
         public async Task<bool> HasPassedAsync(int studentId, int subjectId, CancellationToken ct = default)
         {
-           return await _db.Exams.AnyAsync(e => e.StudentID == studentId
-                                  && e.SubjectID == subjectId
-                                  && e.Grade >= 6, ct);
+            return await _db.Exams.AnyAsync(e => e.StudentID == studentId
+                                   && e.SubjectID == subjectId
+                                   && e.Grade >= 6, ct);
         }
     }
 }
